Initial: Reorganize schedule analysis passes - Add initial ``NodeDurations`` struct.
