+++ crates/transpiler/src/passes/mod.rs
@@ -21,9 +21,7 @@
 //! crate. These are public to be passed to qiskit-pyext and are only used
 //! for building Python submodules.
 
-mod alap_schedule_analysis;
 mod apply_layout;
-mod asap_schedule_analysis;
 mod barrier_before_final_measurement;
 mod basis_translator;
 mod check_map;
@@ -46,6 +44,7 @@ mod optimize_clifford_t;
 mod remove_diagonal_gates_before_measure;
 mod remove_identity_equiv;
 pub mod sabre;
+mod schedule_analysis;
 mod split_2q_unitaries;
 mod substitute_pi4_rotations;
 mod unitary_synthesis;
@@ -53,9 +52,7 @@ mod unroll_3q_or_more;
 pub mod vf2;
 mod wrap_angles;
 
-pub use alap_schedule_analysis::{alap_schedule_analysis_mod, run_alap_schedule_analysis};
 pub use apply_layout::{apply_layout, apply_layout_mod, update_layout};
-pub use asap_schedule_analysis::{asap_schedule_analysis_mod, run_asap_schedule_analysis};
 pub use barrier_before_final_measurement::{
     barrier_before_final_measurements_mod, run_barrier_before_final_measurements,
 };
@@ -90,6 +87,12 @@ pub use remove_diagonal_gates_before_measure::{
     remove_diagonal_gates_before_measure_mod, run_remove_diagonal_before_measure,
 };
 pub use remove_identity_equiv::{remove_identity_equiv_mod, run_remove_identity_equiv};
+pub use schedule_analysis::alap_schedule_analysis::{
+    alap_schedule_analysis_mod, run_alap_schedule_analysis,
+};
+pub use schedule_analysis::asap_schedule_analysis::{
+    asap_schedule_analysis_mod, run_asap_schedule_analysis,
+};
 pub use split_2q_unitaries::{run_split_2q_unitaries, split_2q_unitaries_mod};
 pub use substitute_pi4_rotations::{py_run_substitute_pi4_rotations, substitute_pi4_rotations_mod};
 pub use unitary_synthesis::{
