+++ crates/transpiler/src/passes/schedule_analysis/alap_schedule_analysis.rs
@@ -10,8 +10,8 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
+use super::TimeOps;
 use crate::TranspilerError;
-use crate::passes::alap_schedule_analysis::TimeOps;
 use hashbrown::HashMap;
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
@@ -21,31 +21,35 @@ use qiskit_circuit::operations::{OperationRef, StandardInstruction};
 use qiskit_circuit::{Clbit, Qubit};
 use rustworkx_core::petgraph::prelude::NodeIndex;
 
-pub fn run_asap_schedule_analysis<T: TimeOps>(
+pub fn run_alap_schedule_analysis<T: TimeOps>(
     dag: &DAGCircuit,
     clbit_write_latency: T,
     node_durations: HashMap<NodeIndex, T>,
 ) -> PyResult<HashMap<NodeIndex, T>> {
     if dag.qregs().len() != 1 || !dag.qregs_data().contains_key("q") {
         return Err(TranspilerError::new_err(
-            "ASAP schedule runs on physical circuits only",
+            "ALAP schedule runs on physical circuits only",
         ));
     }
 
     let mut node_start_time: HashMap<NodeIndex, T> = HashMap::new();
-    let mut idle_after: HashMap<Wire, T> = HashMap::new();
+    let mut idle_before: HashMap<Wire, T> = HashMap::new();
 
     let zero = T::zero();
 
     for index in 0..dag.qubits().len() {
-        idle_after.insert(Wire::Qubit(Qubit::new(index)), zero);
+        idle_before.insert(Wire::Qubit(Qubit::new(index)), zero);
     }
 
     for index in 0..dag.clbits().len() {
-        idle_after.insert(Wire::Clbit(Clbit::new(index)), zero);
+        idle_before.insert(Wire::Clbit(Clbit::new(index)), zero);
     }
 
-    for node_index in dag.topological_op_nodes(false) {
+    // Since this is alap scheduling, node is scheduled in reversed topological ordering
+    // and nodes are packed from the very end of the circuit.
+    // The physical meaning of t0 and t1 is flipped here.
+
+    for node_index in dag.topological_op_nodes(true) {
         let op = dag[node_index].unwrap_operation();
 
         let qargs: Vec<Wire> = dag
@@ -61,12 +65,10 @@ pub fn run_asap_schedule_analysis<T: TimeOps>(
             .map(|&c| Wire::Clbit(c))
             .collect();
 
-        let &op_duration = node_durations.get(&node_index).ok_or_else(|| {
-            TranspilerError::new_err(format!(
-                "No duration found for node at index {}",
-                node_index.index()
-            ))
-        })?;
+        let &op_duration = node_durations
+            .get(&node_index)
+            .ok_or_else(|| TranspilerError::new_err("No duration for node"))?;
+
         let op_view = op.op.view();
         let is_gate_or_delay = matches!(
             op_view,
@@ -79,74 +81,67 @@ pub fn run_asap_schedule_analysis<T: TimeOps>(
         // t0: start time of instruction
         // t1: end time of instruction
 
-        let (t0, t1) = if is_gate_or_delay {
-            let t0 = qargs
+        let t1 = if is_gate_or_delay {
+            let &t0 = qargs
                 .iter()
-                .map(|q| idle_after[q])
-                .fold(zero, |acc, x| *T::max(&acc, &x));
-            (t0, t0 + op_duration)
+                .map(|q| idle_before.get(q).unwrap_or(&zero))
+                .fold(&zero, |acc, x| T::max(acc, x));
+            t0 + op_duration
         } else if matches!(
             op_view,
             OperationRef::StandardInstruction(StandardInstruction::Measure)
         ) {
-            // Measure instruction handling is bit tricky due to clbit_write_latency
-            let t0q = qargs
-                .iter()
-                .map(|q| idle_after[q])
-                .fold(zero, |acc, x| *T::max(&acc, &x));
-            let t0c = cargs
-                .iter()
-                .map(|c| idle_after[c])
-                .fold(zero, |acc, x| *T::max(&acc, &x));
-            // Assume following case (t0c > t0q)
-            //
-            //       |t0q
-            // Q ▒▒▒▒░░░░░░░░░░░░
-            // C ▒▒▒▒▒▒▒▒░░░░░░░░
-            //           |t0c
-            //
-            // In this case, there is no actual clbit access until clbit_write_latency.
-            // The node t0 can be push backward by this amount.
-            //
-            //         |t0q' = t0c - clbit_write_latency
-            // Q ▒▒▒▒░░▒▒▒▒▒▒▒▒▒▒
-            // C ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
-            //           |t0c' = t0c
-            //
-            // rather than naively doing
-            //
-            //           |t0q' = t0c
-            // Q ▒▒▒▒░░░░▒▒▒▒▒▒▒▒
-            // C ▒▒▒▒▒▒▒▒░░░▒▒▒▒▒
-            //              |t0c' = t0c + clbut_write_latency
-            let t0 = *T::max(&t0q, &(t0c - clbit_write_latency));
-            let t1 = t0 + op_duration;
-            for clbit in cargs.iter() {
-                idle_after.insert(*clbit, t1);
-            }
-            (t0, t1)
-        } else {
-            // Directives (like Barrier)
-            let t0 = qargs
+            // clbit time is always right (alap) justified
+            let &t0 = qargs
                 .iter()
                 .chain(cargs.iter())
-                .map(|bit| idle_after[bit])
-                .fold(zero, |acc, x| *T::max(&acc, &x));
-            (t0, t0 + op_duration)
+                .map(|bit| idle_before.get(bit).unwrap_or(&zero))
+                .fold(&zero, |acc, x| T::max(acc, x));
+
+            let t1 = t0 + op_duration;
+            //
+            //        |t1 = t0 + duration
+            // Q ░░░░░▒▒▒▒▒▒▒▒▒▒▒
+            // C ░░░░░░░░░▒▒▒▒▒▒▒
+            //            |t0 + (duration - clbit_write_latency)
+
+            for clbit in cargs.iter() {
+                idle_before.insert(*clbit, t0 + op_duration - clbit_write_latency);
+            }
+            t1
+        } else {
+            // Directives (like Barrier)
+            let &t0 = qargs
+                .iter()
+                .chain(cargs.iter())
+                .map(|bit| idle_before.get(bit).unwrap_or(&zero))
+                .fold(&zero, |acc, x| T::max(acc, x));
+            t0 + op_duration
         };
 
         for qubit in qargs {
-            idle_after.insert(qubit, t1);
+            idle_before.insert(qubit, t1);
         }
 
-        node_start_time.insert(node_index, t0);
+        node_start_time.insert(node_index, t1);
     }
 
-    Ok(node_start_time)
+    // Compute maximum instruction available time
+    let circuit_duration = idle_before.values().fold(&zero, |acc, x| T::max(acc, x));
+
+    // Note that ALAP pass is inversely scheduled, thus
+    // t0 is computed by subtracting t1 from the entire circuit duration.
+    let mut result: HashMap<NodeIndex, T> = HashMap::new();
+    for (node_idx, t1) in node_start_time {
+        let final_time = *circuit_duration - t1;
+        result.insert(node_idx, final_time);
+    }
+
+    Ok(result)
 }
 
 #[pyfunction]
-/// Runs the ASAPSchedule analysis pass on dag.
+/// Runs the ALAPSchedule analysis pass on dag.
 ///
 /// Args:
 ///     dag (DAGCircuit): DAG to schedule.
@@ -156,8 +151,8 @@ pub fn run_asap_schedule_analysis<T: TimeOps>(
 /// Returns:
 ///     PyDict: A dictionary mapping each DAGOpNode to its scheduled start time.
 ///
-#[pyo3(name = "asap_schedule_analysis", signature= (dag, clbit_write_latency, node_durations))]
-pub fn py_run_asap_schedule_analysis(
+#[pyo3(name = "alap_schedule_analysis", signature= (dag, clbit_write_latency, node_durations))]
+pub fn py_run_alap_schedule_analysis(
     py: Python,
     dag: &DAGCircuit,
     clbit_write_latency: u64,
@@ -185,7 +180,7 @@ pub fn py_run_asap_schedule_analysis(
             op_durations.insert(node_idx, val);
         }
         let node_start_time =
-            run_asap_schedule_analysis::<u64>(dag, clbit_write_latency, op_durations)?;
+            run_alap_schedule_analysis::<u64>(dag, clbit_write_latency, op_durations)?;
         for (node_idx, t1) in node_start_time {
             let node = dag.get_node(py, node_idx)?;
             py_dict.set_item(node, t1)?;
@@ -204,7 +199,7 @@ pub fn py_run_asap_schedule_analysis(
             op_durations.insert(node_idx, val);
         }
         let node_start_time =
-            run_asap_schedule_analysis::<f64>(dag, clbit_write_latency as f64, op_durations)?;
+            run_alap_schedule_analysis::<f64>(dag, clbit_write_latency as f64, op_durations)?;
         for (node_idx, t1) in node_start_time {
             let node = dag.get_node(py, node_idx)?;
             py_dict.set_item(node, t1)?;
@@ -215,7 +210,7 @@ pub fn py_run_asap_schedule_analysis(
     Ok(py_dict.into())
 }
 
-pub fn asap_schedule_analysis_mod(m: &Bound<PyModule>) -> PyResult<()> {
-    m.add_wrapped(wrap_pyfunction!(py_run_asap_schedule_analysis))?;
+pub fn alap_schedule_analysis_mod(m: &Bound<PyModule>) -> PyResult<()> {
+    m.add_wrapped(wrap_pyfunction!(py_run_alap_schedule_analysis))?;
     Ok(())
 }
